wabot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
aiohttp==3.11.11
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
apscheduler==3.10.4
